Refactor code
- Move actions and reducers to their own folders
