+++ src/actions/index.js
@@ -1,0 +1,9 @@
+export {
+  setSyncProgress,
+  incrementSyncProgress,
+  setSyncError,
+  setSyncTotal,
+  setSyncProgressMessage,
+  setSyncIsSyncing,
+  setSyncCompletionTime,
+} from './SyncActions';
