+++ src/navigation/index.js
@@ -2,7 +2,8 @@
  * mSupply Mobile
  * Sustainable Solutions (NZ) Ltd. 2019
  */
+import ReduxNavigator from './ReduxNavigation';
+import AppNavigator from './Navigator';
 
-export AppNavigator from './Navigator';
-export navigationReducer from './reducer';
 export { getCurrentRouteName, getCurrentParams } from './selectors';
+export { AppNavigator, ReduxNavigator };
