+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import NavigationReducer from './NavigationReducer';
+import SyncReducer from './SyncReducer';
+
+export default combineReducers({
+  nav: NavigationReducer,
+  sync: SyncReducer,
+});
