+++ src/sync/index.js
@@ -1,4 +1,3 @@
 export { PostSyncProcessor } from './PostSyncProcessor';
-export syncReducer from './reducer';
 export { Synchroniser } from './Synchroniser';
 export { SyncModal } from './SyncModal';
